+++ package.json
@@ -27,22 +27,21 @@
     "url": "https://github.com/checkdigit/eslint-plugin/issues"
   },
   "homepage": "https://github.com/checkdigit/eslint-plugin#readme",
-  "dependencies": {},
   "peerDependencies": {
     "eslint": ">=7"
   },
   "devDependencies": {
     "@checkdigit/prettier-config": "^1.0.2",
     "@checkdigit/publish": "^1.0.19",
-    "@checkdigit/typescript-config": "^1.0.5",
+    "@checkdigit/typescript-config": "^2.1.0",
     "@types/eslint": "^7.2.11",
     "@types/jest": "^26.0.23",
     "@types/node": "^15.6.1",
     "eslint": "^7.27.0",
-    "jest": "^26.6.3",
+    "jest": "^27.0.1",
     "prettier": "^2.3.0",
-    "ts-jest": "^26.5.6",
-    "typescript": "^4.2.4"
+    "ts-jest": "^27.0.0",
+    "typescript": "^4.3.2"
   },
   "jest": {
     "collectCoverage": true,
